requirements.txt
flask==2.3.3
werkzeug==2.3.7
python-dotenv==1.0.0
mistralai==0.0.7
numpy==1.24.3
pandas==2.1.4
reportlab==4.0.8
nltk==3.8.1
textblob==0.17.1
opencv-python==4.8.1.78
torch==2.2.0
transformers==4.36.2
ffmpeg-python==0.2.0
sounddevice==0.4.6
openai-whisper==20231117
PyPDF2==3.0.1
librosa==0.10.1
soundfile==0.12.1
spacy==3.7.2
tensorflow==2.13.0
deepface==0.0.79
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl 